Provided a debugging flag that suppresses enum definitions for whan I'm working on other stuff.
Added a lot of type-renaming stuff from the original readspec.rkt.
and provided some of my own as a stopgap until I figure out whar really has to happen here.

made al output operate in a uniform way with fprintf.
Also provided a second name 'anomaly' for the output stream;
anomaly is intended to contain messages, especially errors and incompleteness reports.

Removed obsolete debug trace messages.

Process commands now, but incompletely.
Cleaned up parameter processing code to make it easier to extend.  It still has a long way to go.

Used 'when' instead of 'if when this is appropriate.